torch-server/requirements.txt
chardet==3.0.4
Click==7.0
Flask==1.0.2
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
numpy==1.16.2
Pillow==5.4.1
requests==2.21.0
six==1.12.0
torch==1.0.1
torchvision==0.2.2
urllib3==1.24.1
Werkzeug==0.15.1
